+++ src/object/setProp.ts
@@ -1,17 +1,18 @@
-import setPropByList from './setPropByList'
-
+import config from '../../config'
 /**
- * 根据a.b字符串设置属性 // 此函数认为targetData和prop必然存在，不做单独校验
+ * 根据字符串设置属性
  * @param {*} targetData 对应对象
- * @param {string} prop 属性字符串a.b,,父属性不存在时会创建对象
+ * @param {string} prop 属性字符串
  * @param {*} value 属性值
  * @param {boolean} [useSetData] 为真时通过setData进行赋值操作,主要针对框架中直接赋值无法响应的操作
  * @returns {boolean} 设置是否成功
  */
 function setProp(targetData: Record<PropertyKey, any>, prop: string, value: unknown, useSetData?: boolean) {
-  const propList = prop.indexOf('.') > -1 ? prop.split('.') : [prop]
-  setPropByList(targetData, propList, value, useSetData)
-  return true
+  if (!useSetData) {
+    targetData[prop] = value
+  } else {
+    config.object.setData(targetData, prop, value)
+  }
 }
 
 export default setProp